pyproject.toml
[project]
name = "nonebot-plugin-pypistats"
version = "0.1.4"
description = "Default template for PDM package"
authors = [
    {name = "zhongwen-4", email = "2401128923@qq.com"},
]
dependencies = [
    "nonebot2>=2.4.0",
    "nonebot-plugin-alconna>=0.53.1",
    "httpx>=0.27.2",
    "nonebot-plugin-send-anything-anywhere>=0.7.1",
    "pyecharts>=2.0.7",
    "nonebot-plugin-htmlrender>=0.4.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
nb_run = 'cd ../ && cd ../ && cd ../ && nb run --reload'

[tool.ruff.lint]
ignore = ["E402"]
